[SYCL][Matrix] Add explicit cast for the conditional operator to avoid unexpected behavior (intel/llvm-test-suite#727)

+++ SYCL/Matrix/element_wise_all_ops_half.cpp
@@ -41,10 +41,9 @@ public:
 };
 
 template <typename T, size_t M, size_t N>
-void assert_ops_ref(/*const T &C*/ accessor<T, 2, access::mode::read,
-                                            access::target::host_buffer>
-                        C,
-                    const float ref) {
+void assert_ops_ref(
+    accessor<T, 2, access::mode::read, access::target::host_buffer> C,
+    const float ref) {
   for (size_t i = 0; i < M; i++)
     for (size_t j = 0; j < N; j++) {
       auto diff = C[i][j] - ref;
@@ -204,7 +203,8 @@ void matrix_verify_logic(queue q, big_matrix<T, M, N> &A, nd_range<2> &r,
          if (wi_slice_a[i]) {
            if (wi_slice_a[i] > 2.0 || wi_slice_a[i] >= 2.0 ||
                wi_slice_a[i] < 2.0 || wi_slice_a[i] <= 2.0) {
-             T val = (wi_slice_a[i] != 2.0) ? wi_slice_a[i] : 2.0;
+             T val = (wi_slice_a[i] != 2.0) ? wi_slice_a[i]
+                                            : static_cast<half>(2.0);
              val--;
              val++;
              if (wi_slice_a[i] == 2.0) {
